pyproject.toml
[project]
name = "tarkovdiscbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "hikari>=2.5.0",
    "hikari-arc[cron]>=2.1.1",
    "hikari-miru>=4.2.0",
    "tomlkit>=0.14.0",
]

[project.optional-dependencies]
linux = [
    "uvloop>=0.22.1",
]
lint = [
    "ruff>=0.14.13",
]
dev = [
    "pytest>=9.0.2",
]
